+++ ta4j-core/src/main/java/org/ta4j/core/num/Num.java
@@ -24,6 +24,9 @@
 package org.ta4j.core.num;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+import java.math.MathContext;
+import java.math.RoundingMode;
 import java.util.function.Function;
 
 /**
@@ -202,6 +205,18 @@ public interface Num extends Comparable<Num>, Serializable {
      */
     default Num numOf(Number value){
         return function().apply(value);
+    }
+
+    /**
+     * Transforms a {@link String} into a new Num instance of this with a precision
+     * <code>Num</code> implementation
+     * @param value the String to transform
+     * @param precision the precision
+     * @return the corresponding Num implementation of the <code>value</code>
+     */
+    default Num numOf(String string, int precision) {
+        MathContext mathContext = new MathContext(precision, RoundingMode.HALF_UP);
+        return this.numOf(new BigDecimal(string, mathContext));
     }
 
     /**